requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.2.1
Django==2.2
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-json-field==0.5.7
django-jsonfield==1.4.1
django-user-agents==0.4.0
idna==2.10
jwt==1.1.0
Pillow==8.1.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.2
user-agents==2.2.0
